# api/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Shopmon API
  description: Shopware monitoring application API
  version: 1.0.0
servers:
  - url: /api
    description: API base path
...
security:
  - bearerAuth: []
paths:
...
  /notifications:
    get:
      operationId: getNotifications
      tags: [Notifications]
      summary: Get all notifications for the current user
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
        "401":
          $ref: "#/components/responses/Unauthorized"
    delete:
      operationId: deleteAllNotifications
      tags: [Notifications]
      summary: Delete all notifications
      responses:
        "204":
          description: All notifications deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [Notifications]
      summary: Delete a single notification
      parameters:
        - $ref: "#/components/parameters/NotificationId"
      responses:
        "204":
          description: Notification deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /notifications/mark-read:
    post:
      operationId: markNotificationsRead
      tags: [Notifications]
      summary: Mark all notifications as read
      responses:
        "204":
          description: All notifications marked as read
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ──────────────────────────────────────────────
  # Shops
  # ──────────────────────────────────────────────
...
  /shops/{shopId}/deployments:
    get:
      operationId: getDeployments
      tags: [Deployments]
      summary: List deployments for a shop
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /shops/{shopId}/deployments/{deploymentId}:
    get:
      operationId: getDeployment
      tags: [Deployments]
      summary: Get deployment details with output
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - $ref: "#/components/parameters/DeploymentId"
      responses:
        "200":
          description: Deployment details with output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteDeployment
      tags: [Deployments]
      summary: Delete a deployment
      parameters:
        - $ref: "#/components/parameters/ShopId"
        - $ref: "#/components/parameters/DeploymentId"
      responses:
        "204":
          description: Deployment deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ──────────────────────────────────────────────
  # CLI
  # ──────────────────────────────────────────────
  /cli/deployments:
    post:
      operationId: createCliDeployment
      tags: [CLI]
      summary: Create a deployment via CLI
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCliDeploymentRequest"
      responses:
        "201":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCliDeploymentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          $ref: "#/components/responses/ValidationError"

  # ──────────────────────────────────────────────
  # SSO
  # ──────────────────────────────────────────────
...
  /info/extension-compatibility:
    post:
      operationId: checkExtensionCompatibility
      tags: [Info]
      summary: Check extension compatibility between Shopware versions
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExtensionCompatibilityRequest"
      responses:
        "200":
          description: Compatibility results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExtensionCompatibilityResult"
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication

  # ──────────────────────────────────────────────
  # Parameters
  # ──────────────────────────────────────────────
  parameters:
...
    ShopId:
      name: shopId
      in: path
      required: true
      schema:
        type: integer
      description: Shop ID
...
    DeploymentId:
      name: deploymentId
      in: path
      required: true
      schema:
        type: integer
      description: Deployment ID
...
    NotificationId:
      name: id
      in: path
      required: true
      schema:
        type: integer
      description: Notification ID

  # ──────────────────────────────────────────────
  # Responses
  # ──────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  # ──────────────────────────────────────────────
  # Schemas
  # ──────────────────────────────────────────────
  schemas:
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string

    # ── Account ──────────────────────────────────
...
    Notification:
      type: object
      required:
        - id
        - userId
        - key
        - level
        - title
        - message
        - link
        - read
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: string
        key:
          type: string
        level:
          type: string
        title:
          type: string
        message:
          type: string
        link:
          allOf:
            - $ref: "#/components/schemas/NotificationLink"
          nullable: true
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
    NotificationLink:
      type: object
      required:
        - url
        - label
      properties:
        url:
          type: string
        label:
          type: string

    # ── Shops ────────────────────────────────────
...
    Deployment:
      type: object
      required:
        - id
        - name
        - command
        - returnCode
        - startDate
        - endDate
        - executionTime
        - reference
        - createdAt
      properties:
        id:
          type: integer
        name:
          type: string
          nullable: true
        command:
          type: string
        returnCode:
          type: integer
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
        executionTime:
          type: number
          description: Execution time in seconds
        reference:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    DeploymentDetail:
      allOf:
        - $ref: "#/components/schemas/Deployment"
        - type: object
          properties:
            output:
              type: string
              nullable: true
    CreateCliDeploymentRequest:
      type: object
      required:
        - shop_id
        - command
        - return_code
        - start_date
        - end_date
        - execution_time
      properties:
        shop_id:
          type: integer
        command:
          type: string
        return_code:
          type: integer
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
        execution_time:
          type: number
        composer:
          type: string
          nullable: true
        reference:
          type: string
          nullable: true
        name:
          type: string
          nullable: true
    CreateCliDeploymentResponse:
      type: object
      required:
        - success
        - name
        - deployment_id
        - url
        - upload_url
      properties:
        success:
          type: boolean
        name:
          type: string
        deployment_id:
          type: integer
        url:
          type: string
        upload_url:
          type: string

    # ── SSO ──────────────────────────────────────
...
    ExtensionCompatibilityRequest:
      type: object
      required:
        - currentVersion
        - futureVersion
        - extensions
      properties:
        currentVersion:
          type: string
        futureVersion:
          type: string
        extensions:
          type: array
          items:
            type: object
            required:
              - name
              - version
            properties:
              name:
                type: string
              version:
                type: string
    ExtensionCompatibilityResult:
      type: object
      required:
        - name
        - label
        - iconPath
        - status
      properties:
        name:
          type: string
        label:
          type: string
        iconPath:
          type: string
          nullable: true
        status:
          type: object
          required:
            - name
            - label
            - type
          properties:
            name:
              type: string
            label:
              type: string
            type:
              type: string

    # ── Auth ────────────────────────────────────
...